+++ execution_engine/src/core/runtime_context/mod.rs
@@ -330,8 +330,12 @@ where
                 error!("should not remove the chainspec registry key");
                 Err(Error::RemoveKeyFailure(RemoveKeyFailure::PermissionDenied))
             }
-            Key::BlockEffectsRootHash { block_height: _ } => {
+            Key::BlockEffectsRootHash { .. } => {
                 error!("should not remove the execution results root hash key");
+                Err(Error::RemoveKeyFailure(RemoveKeyFailure::PermissionDenied))
+            }
+            Key::DeployApprovalsRootHash { .. } => {
+                error!("should not remove the deploy approvals root hash key");
                 Err(Error::RemoveKeyFailure(RemoveKeyFailure::PermissionDenied))
             }
         }
@@ -790,7 +794,8 @@ where
             Key::Dictionary(_) => true,
             Key::SystemContractRegistry => true,
             Key::ChainspecRegistry => true,
-            Key::BlockEffectsRootHash { block_height: _ } => true,
+            Key::BlockEffectsRootHash { .. } => true,
+            Key::DeployApprovalsRootHash { .. } => true,
         }
     }
 
@@ -813,7 +818,8 @@ where
             }
             Key::SystemContractRegistry => false,
             Key::ChainspecRegistry => false,
-            Key::BlockEffectsRootHash { block_height: _ } => false,
+            Key::BlockEffectsRootHash { .. } => false,
+            Key::DeployApprovalsRootHash { .. } => false,
         }
     }
 
@@ -836,7 +842,8 @@ where
             }
             Key::SystemContractRegistry => false,
             Key::ChainspecRegistry => false,
-            Key::BlockEffectsRootHash { block_height: _ } => false,
+            Key::BlockEffectsRootHash { .. } => false,
+            Key::DeployApprovalsRootHash { .. } => false,
         }
     }