+++ node/src/components/contract_runtime/operations.rs
@@ -1,4 +1,4 @@
-use std::{collections::BTreeMap, iter, sync::Arc, time::Instant};
+use std::{collections::BTreeMap, sync::Arc, time::Instant};
 
 use itertools::Itertools;
 use tracing::{debug, trace, warn};
@@ -65,6 +65,7 @@ pub fn execute_finalized_block(
     // Run any deploys that must be executed
     let block_time = finalized_block.timestamp().millis();
     let start = Instant::now();
+    let deploy_approvals_root_hash = compute_approvals_root_hash(&deploys, &transfers)?;
 
     // Create a new EngineState that reads from LMDB but only caches changes in memory.
     let scratch_state = engine_state.get_scratch_engine_state();
@@ -81,7 +82,7 @@ pub fn execute_finalized_block(
         );
 
         // TODO: this is currently working coincidentally because we are passing only one
-        // deploy_item per exec. The execution results coming back from the ee lacks the
+        // deploy_item per exec. The execution results coming back from the EE lack the
         // mapping between deploy_hash and execution result, and this outer logic is
         // enriching it with the deploy hash. If we were passing multiple deploys per exec
         // the relation between the deploy and the execution results would be lost.
@@ -105,21 +106,24 @@ pub fn execute_finalized_block(
     )?;
     let block_height = finalized_block.height();
 
-    let execution_results_root_hash_effect = iter::once((
+    let mut effects = AdditiveMap::new();
+    let _ = effects.insert(
+        Key::DeployApprovalsRootHash { block_height },
+        Transform::Write(
+            CLValue::from_t(deploy_approvals_root_hash)
+                .map_err(BlockCreationError::CLValue)?
+                .into(),
+        ),
+    );
+    let _ = effects.insert(
         Key::BlockEffectsRootHash { block_height },
         Transform::Write(
             CLValue::from_t(execution_results_root_hash)
                 .map_err(BlockCreationError::CLValue)?
                 .into(),
         ),
-    ))
-    .collect();
-
-    scratch_state.apply_effect(
-        CorrelationId::new(),
-        state_root_hash,
-        execution_results_root_hash_effect,
-    )?;
+    );
+    scratch_state.apply_effect(CorrelationId::new(), state_root_hash, effects)?;
 
     if let Some(metrics) = metrics.as_ref() {
         metrics.exec_block.observe(start.elapsed().as_secs_f64());
@@ -202,19 +206,6 @@ pub fn execute_finalized_block(
         execution_results,
         maybe_step_effect_and_upcoming_era_validators,
     })
-}
-
-/// Computes the root hash for a Merkle tree constructed from the root hashes
-fn compute_execution_results_root_hash<'a>(
-    results: &mut impl Iterator<Item = &'a ExecutionResult>,
-) -> Result<Digest, BlockCreationError> {
-    let mut execution_results = vec![];
-    for result in results {
-        execution_results.push(Digest::hash(
-            &result.to_bytes().map_err(BlockCreationError::BytesRepr)?,
-        ));
-    }
-    Ok(Digest::hash_merkle_tree(execution_results))
 }
 
 /// Commits the execution effects.
@@ -406,4 +397,33 @@ where
     }
     trace!(?result, "step response");
     result
+}
+
+/// Computes the root hash for a Merkle tree constructed from the hashes of execution results.
+fn compute_execution_results_root_hash<'a>(
+    results: &mut impl Iterator<Item = &'a ExecutionResult>,
+) -> Result<Digest, BlockCreationError> {
+    let mut execution_results = vec![];
+    for result in results {
+        execution_results.push(Digest::hash(
+            &result.to_bytes().map_err(BlockCreationError::BytesRepr)?,
+        ));
+    }
+    Ok(Digest::hash_merkle_tree(execution_results))
+}
+
+/// Computes the root hash for a Merkle tree constructed from the hashes of deploy approvals.
+fn compute_approvals_root_hash(
+    deploys: &[Deploy],
+    transfers: &[Deploy],
+) -> Result<Digest, BlockCreationError> {
+    let mut approval_hashes = vec![];
+    for deploy in deploys.iter().chain(transfers) {
+        let bytes = deploy
+            .approvals()
+            .to_bytes()
+            .map_err(BlockCreationError::BytesRepr)?;
+        approval_hashes.push(Digest::hash(bytes));
+    }
+    Ok(Digest::hash_merkle_tree(approval_hashes))
 }