+++ types/src/cl_value.rs
@@ -19,7 +19,7 @@ use crate::{
 mod jsonrepr;
 
 /// Error while converting a [`CLValue`] into a given type.
-#[derive(PartialEq, Eq, Clone, Debug, Serialize)]
+#[derive(PartialEq, Eq, Clone, Debug, Serialize, Deserialize)]
 #[cfg_attr(feature = "datasize", derive(DataSize))]
 pub struct CLTypeMismatch {
     /// The [`CLType`] into which the `CLValue` was being converted.
@@ -40,7 +40,7 @@ impl Display for CLTypeMismatch {
 }
 
 /// Error relating to [`CLValue`] operations.
-#[derive(PartialEq, Eq, Clone, Debug, Serialize)]
+#[derive(PartialEq, Eq, Clone, Debug, Serialize, Deserialize)]
 #[cfg_attr(feature = "datasize", derive(DataSize))]
 pub enum CLValueError {
     /// An error while serializing or deserializing the underlying data.